+++ runtime/src/main/java/com/xatkit/plugins/core/platform/action/StoreValue.java
@@ -6,13 +6,19 @@ import com.xatkit.plugins.core.platform.CorePlatform;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
+import java.util.Map;
 
 import static fr.inria.atlanmod.commons.Preconditions.checkNotNull;
 
 /**
- * Puts the given {@code key=value} pair in the {@link CorePlatform}'s {@code store}.
+ * Puts the given {@code key=value} pair in the provided {@code collection}.
  */
 public class StoreValue extends RuntimeAction<CorePlatform> {
+
+    /**
+     * The collection associated to the value to store.
+     */
+    private String collectionName;
 
     /**
      * The key associated to the value to store.
@@ -30,16 +36,29 @@ public class StoreValue extends RuntimeAction<CorePlatform> {
      *
      * @param runtimePlatform the {@link CorePlatform} containing this action
      * @param session         the {@link XatkitSession} associated to this action
+     * @param collectionName  the name of the collection associated to the value to store
      * @param key             the key associated to the value to store
      * @param value           the value to store
-     * @throws NullPointerException if the provided {@code key} is {@code null}
+     * @throws NullPointerException if the provided {@code collection} or {@code key} is {@code null}
      */
-    public StoreValue(CorePlatform runtimePlatform, XatkitSession session, @Nonnull String key,
-                      @Nullable Object value) {
+    public StoreValue(CorePlatform runtimePlatform, XatkitSession session,
+                      @Nonnull String collectionName, @Nonnull String key, @Nullable Object value) {
         super(runtimePlatform, session);
+        checkNotNull(collectionName, "Cannot construct a %s from the provided collection name %s",
+                this.getClass().getSimpleName(), collectionName);
         checkNotNull(key, "Cannot construct a %s from the provided key %s", this.getClass().getSimpleName(), key);
+        this.collectionName = collectionName;
         this.key = key;
         this.value = value;
+    }
+
+    /**
+     * Returns the name of the collection associated to the value to store.
+     *
+     * @return the name of the collection associated to the value to store
+     */
+    public String getCollectionName() {
+        return this.collectionName;
     }
 
     /**
@@ -61,7 +80,7 @@ public class StoreValue extends RuntimeAction<CorePlatform> {
     }
 
     /**
-     * Puts the given {@code key=value} pair in the {@link CorePlatform}'s {@code store}.
+     * Puts the given {@code key=value} pair in the provided {@code collection}.
      * <p>
      * This method overwrites any value already associated to the given {@code key}.
      *
@@ -69,7 +88,8 @@ public class StoreValue extends RuntimeAction<CorePlatform> {
      */
     @Override
     protected Object compute() {
-        this.runtimePlatform.getStore().put(key, value);
+        Map<String, Object> store = this.runtimePlatform.getOrCreateCollection(collectionName);
+        store.put(key, value);
         return value;
     }
 }