+++ runtime/src/test/java/com/xatkit/plugins/core/platform/action/GetValueTest.java
@@ -15,67 +15,93 @@ public class GetValueTest extends AbstractActionTest<GetValue, CorePlatform> {
     }
 
     @Test(expected = NullPointerException.class)
+    public void constructNullCollection() {
+        action = new GetValue(platform, emptySession, null, "key");
+    }
+
+
+    @Test(expected = NullPointerException.class)
     public void constructNullKey() {
-        action = new GetValue(platform, emptySession, null);
+        action = new GetValue(platform, emptySession, "collection", null);
     }
 
     @Test
-    public void constructValidKey() {
-        action = new GetValue(platform, emptySession, "key");
-        assertThatFieldsAreSet(action, "key", null);
+    public void constructValidCollectionAndKey() {
+        action = new GetValue(platform, emptySession, "collection", "key");
+        assertThatFieldsAreSet(action, "collection", "key", null);
     }
 
     @Test
-    public void constructValidKeyNullDefaultValue() {
-        action = new GetValue(platform, emptySession, "key", null);
-        assertThatFieldsAreSet(action, "key", null);
+    public void constructValidCollectionAndKeyNullDefaultValue() {
+        action = new GetValue(platform, emptySession, "collection", "key", null);
+        assertThatFieldsAreSet(action, "collection", "key", null);
     }
 
     @Test
-    public void constructValidKeyNotNullDefaultValue() {
-        action = new GetValue(platform, emptySession, "key", "default");
-        assertThatFieldsAreSet(action, "key", "default");
+    public void constructValidCollectionAndKeyNotNullDefaultValue() {
+        action = new GetValue(platform, emptySession, "collection", "key", "default");
+        assertThatFieldsAreSet(action, "collection", "key", "default");
     }
 
     @Test
-    public void computeValueNotInStoreNoDefaultValue() {
-        action = new GetValue(platform, emptySession, "key");
+    public void computeCollectionDoesNotExistNoDefaultValue() {
+        action = new GetValue(platform, emptySession, "collection", "key");
         Object result = action.compute();
         assertThat(result).as("Result is null").isNull();
     }
 
     @Test
-    public void computeValueNotInStoreDefaultValue() {
-        action = new GetValue(platform, emptySession, "key", "default");
+    public void computeCollectionDoesNotExistDefaultValue() {
+        action = new GetValue(platform, emptySession, "collection", "key", "default");
         Object result = action.compute();
         assertThat(result).as("Result is default value").isEqualTo("default");
     }
 
     @Test
-    public void computeValueInStoreNoDefaultValue() {
-        platform.getStore().put("key", "value");
-        action = new GetValue(platform, emptySession, "key");
-        Object result = action.compute();
-        assertThat(result).as("Valid result").isEqualTo("value");
-    }
-
-    @Test
-    public void computeValueInStoreDefaultValue() {
-        platform.getStore().put("key", "value");
-        action = new GetValue(platform, emptySession, "key", "default");
-        Object result = action.compute();
-        assertThat(result).as("Valid result").isEqualTo("value");
-    }
-
-    @Test
-    public void computeNullValueInStoreDefaultValue() {
-        platform.getStore().put("key", null);
-        action = new GetValue(platform, emptySession, "key", "default");
+    public void computeValueNotInCollectionNoDefaultValue() {
+        /*
+         * Create an empty collection.
+         */
+        platform.getOrCreateCollection("collection");
+        action = new GetValue(platform, emptySession, "collection", "key");
         Object result = action.compute();
         assertThat(result).as("Result is null").isNull();
     }
 
-    private void assertThatFieldsAreSet(GetValue action, String key, Object defaultValue) {
+    @Test
+    public void computeValueNotInCollectionDefaultValue() {
+        platform.getOrCreateCollection("collection");
+        action = new GetValue(platform, emptySession, "collection", "key", "default");
+        Object result = action.compute();
+        assertThat(result).as("Result is default value").isEqualTo("default");
+    }
+
+    @Test
+    public void computeValueInCollectionNoDefaultValue() {
+        platform.getOrCreateCollection("collection").put("key", "value");
+        action = new GetValue(platform, emptySession, "collection", "key");
+        Object result = action.compute();
+        assertThat(result).as("Valid result").isEqualTo("value");
+    }
+
+    @Test
+    public void computeValueInCollectionDefaultValue() {
+        platform.getOrCreateCollection("collection").put("key", "value");
+        action = new GetValue(platform, emptySession, "collection", "key", "default");
+        Object result = action.compute();
+        assertThat(result).as("Valid result").isEqualTo("value");
+    }
+
+    @Test
+    public void computeNullValueInCollectionDefaultValue() {
+        platform.getOrCreateCollection("collection").put("key", null);
+        action = new GetValue(platform, emptySession, "collection", "key", "default");
+        Object result = action.compute();
+        assertThat(result).as("Result is null").isNull();
+    }
+
+    private void assertThatFieldsAreSet(GetValue action, String collection, String key, Object defaultValue) {
+        assertThat(action.getCollectionName()).as("Collection field correctly set").isEqualTo(collection);
         assertThat(action.getKey()).as("Key field correctly set").isEqualTo(key);
         assertThat(action.getDefaultValue()).as("Default value correctly set").isEqualTo(defaultValue);
     }